Delete non-existent resource should return 404 Not Found

+++ lib/rack_dav/controller.rb
@@ -71,6 +71,8 @@ module RackDAV
     end
 
     def delete
+      raise NotFound if not resource.exist?
+
       delete_recursive(resource, errors = [])
 
       if errors.empty?
